moderation_service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
awscli==1.44.44
boto3==1.42.54
botocore==1.42.54
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
docutils==0.19
dotenv==0.9.9
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
flatbuffers==25.12.19
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
onnxruntime==1.19.2
packaging==26.0
pillow==11.3.0
protobuf==6.33.5
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rsa==4.7.2
s3transfer==0.16.0
safetensors==0.7.0
scipy==1.13.1
starlette==0.49.3
sympy==1.14.0
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn==0.39.0
tritonclient[grpc]>=2.40.0